// backend/src/types/EventsMap.ts
export type Response = {
	isSuccess: boolean;
	message: Message | ErrorMessage;
	res?: Object;
};

export type Callback = (response: Response) => void;

export enum Message {
	USER_SUCCESSFULLY_CREATED = 'User successfully created',
	USER_SUCCESSFULLY_UPDATED = 'User successfully edited',
	STUDENT_SUCCESSFULLY_ADDED_TO_CLASS = 'Student successfully added to class',
	STUDENT_SUCCESSFULLY_ADDED_TO_COURSE = 'Student successfully added to course',
	STUDENT_SUCCESSFULLY_CREATED = 'Student successfully created',
	STUDENT_SUCCESSFULLY_FETCHED = 'Student successfully fetched',
	STUDENTS_SUCCESSFULLY_FETCHED = 'Students successfully fetched',
	STUDENT_SUCCESSFULLY_REMOVED_FROM_CLASS = 'Student successfully removed from class',
	TEACHER_SUCCESSFULLY_CREATED = 'Teacher successfully created',
	TEACHER_SUCCESSFULLY_FETCHED = 'Teacher successfully fetched',
	TEACHER_SUCCESSFULLY_ADDED_TO_COURSE = 'Teacher successfully added to course',
	TEACHER_SUCCESSFULLY_REMOVED_FROM_COURSE = 'Teacher successfully removed from course',
	CLASS_SUCCESSFULLY_CREATED = 'Class successfully created',
	CLASS_SUCCESSFULLY_FETCHED = 'Class successfully fetched',
	COURSE_SUCCESSFULLY_CREATED = 'Course successfully created',
	COURSE_SUCCESSFULLY_UPDATED = 'Course successfully updated',
	COURSE_SUCCESSFULLY_FETCHED = 'Course successfully fetched',
	CLASS_SUCCESSFULLY_ADDED_TO_COURSE = 'Class successfully added to course',
	CLASS_SUCCESSFULLY_REMOVED_FROM_COURSE = 'Class successfully removed to course',
	GRADE_SUCCESSFULLY_CREATED = 'Grade created successfully',
	GRADE_SUCCESSFULLY_UPDATED = 'Grade updated successfully',
	GRADE_SUCCESSFULLY_DELETED = 'Grade deleted successfully',
	EXAM_SUCCESSFULLY_CREATED = 'Exam created successfully',
	EXAM_SUCCESSFULLY_FETCHED = 'Exam fetched successfully',
	EXAM_SUCCESSFULLY_UPDATED = 'Exam updated successfully',
	EXAM_SUCCESSFULLY_DELETED = 'Exam deleted successfully',
	ROOM_SUCCESSFULLY_CREATED = 'Room created successfully',
	ROOM_SUCCESSFULLY_UPDATED = 'Room updated successfully',
	ROOM_SUCCESSFULLY_DELETED = 'Room deleted successfully',
	ROOM_SUCCESSFULLY_FETCHED = 'Room fetched successfully',
	ROOMS_SUCCESSFULLY_FETCHED = 'Rooms fetched successfully',
}

export enum ErrorMessage {
	PASSWORD_IS_INCORRECT = 'Password is wrong',
	OLD_PASSWORD_IS_INCORRECT = 'Old password is wrong',
	DAY_OR_TIME_NOT_PROVIDED = 'Day or time not provided',
	COULD_NOT_UPDATE_PASSWORD = 'Could not update password',
	USER_COULD_NOT_BE_FOUND = 'User could not be found',
	USER_COULD_NOT_BE_CREATED = 'User could not be created',
	USER_ALREADY_EXISTS = 'User already exists',
	USER_DOES_NOT_EXIST = 'User does not exist',
	USER_COULD_NOT_BE_UPDATED = 'User could not be edited',
	STUDENT_COULD_NOT_BE_FOUND = 'Student could not be found',
	STUDENT_COULD_NOT_BE_CREATED = 'Student could not be created',
	STUDENT_ALREADY_EXISTS = 'Student already exists',
	STUDENT_ALREADY_IN_CLASS = 'Student already in class',
	STUDENT_ALREADY_IN_COURSE = 'Student already assigned to course',
	STUDENT_IS_NOT_IN_CLASS = 'Student is not in class',
	STUDENT_IS_NOT_IN_COURSE = 'Student is not in course',
	STUDENT_COULD_NOT_BE_ADDED_TO_CLASS = 'Student could not be added to class',
	STUDENT_COULD_NOT_BE_ADDED_TO_COURSE = 'Student could not be added to course',
	STUDENT_IS_ALREADY_ASSIGNED_TO_CLASS = 'Student is already assigned to class',
	STUDENT_COULD_NOT_BE_REMOVED_FROM_CLASS = 'Student could not be removed from class',
	FETCHING_STUDENT_FAILED = 'Fetching student failed',
	TEACHER_COULD_NOT_BE_FOUND = 'Teacher could not be found',
	TEACHER_DOES_NOT_EXIST = 'Teacher does not exist',
	TEACHER_COULD_NOT_BE_CREATED = 'Teacher could not be created',
	TEACHER_COULD_NOT_BE_FETCHED = 'Teacher could not be fetched',
	TEACHER_COULD_NOT_BE_ADDED_TO_COURSE = 'Teacher could not be added to course',
	TEACHER_COULD_NOT_BE_REMOVED_FROM_COURSE = 'Teacher could not be removed from course',
	TEACHER_COULD_NOT_BE_DELETED = 'Teacher could not be deleted',
	TEACHER_ALREADY_IN_COURSE = 'Teacher already in course',
	TEACHER_ALREADY_IN_CLASS = 'Teacher already in class',
	TEACHER_ALREADY_EXISTS = 'Teacher already exists',
	TEACHER_IS_NOT_IN_COURSE = 'Teacher is not in course',
	TEACHER_SUCCESSFULLY_ADDED_TO_CLASS = 'Teacher successfully added to class',
	TEACHER_COULD_NOT_BE_ADDED_TO_CLASS = 'Teacher could not be added to class',
	FETCHING_TEACHER_FAILED = 'Fetching teacher failed',
	CLASS_COULD_NOT_BE_FOUND = 'Class could not be found',
	CLASS_ALREADY_EXISTS = 'Class already exists',
	CLASS_COULD_NOT_BE_CREATED = 'Class could not be created',
	CLASS_COULD_NOT_BE_UPDATED = 'Class could not be updated',
	CLASS_COULD_NOT_BE_DELETED = 'Class could not be deleted',
	CLASS_COULD_NOT_BE_ADDED_TO_COURSE = 'Class could not be added to course',
	CLASS_COULD_NOT_BE_REMOVED_FROM_COURSE = 'Class could not be removed from course',
	CLASS_IS_NOT_ASSIGNED_TO_COURSE = 'Class is not assigned to course',
	CLASS_IS_ALREADY_ASSIGNED_TO_COURSE = 'Class is already assigned to course',
	CLASSES_COULD_NOT_BE_RETURNED = 'Classes could not be fetched',
	COURSE_COULD_NOT_BE_FOUND = 'Course could not be found',
	COURSE_COULD_NOT_BE_CREATED = 'Course could not be created',
	COURSE_COULD_NOT_BE_UPDATED = 'Course could not be updated',
	COURSE_COULD_NOT_BE_DELETED = 'Course could not be deleted',
	COURSE_DOES_NOT_EXIST = 'Course does not exist',
	COURSE_ALREADY_EXISTS = 'Course already exists',
	COURSE_IS_NOT_ASSIGNED_TO_CLASS = 'Course is not assigned to class',
	COURSE_COULD_NOT_BE_ADDED_TO_ROOM = 'Course could not be added to room',
	COURSE_IS_ALREADY_ASSIGNED_TO_ROOM = 'Course is already assigned to room',
	GRADE_COULD_NOT_BE_FOUND = 'Grade could not be found',
	GRADE_COULD_NOT_BE_CREATED = 'Grade could not be created',
	GRADE_COULD_NOT_BE_UPDATED = 'Grade could not be updated',
	GRADE_COULD_NOT_BE_DELETED = 'Grade could not be deleted',
	EXAM_ALREADY_EXISTS = 'Exam already exists',
	EXAM_COULD_NOT_BE_CREATED = 'Exam could not be created',
	EXAM_COULD_NOT_BE_UPDATED = 'Exam could not be updated',
	EXAM_COULD_NOT_BE_DELETED = 'Exam could not be deleted',
	EXAM_COULD_NOT_BE_FOUND = 'Exam could not be found',
	ROOMS_COULD_NOT_BE_FOUND = 'Rooms could not be found',
	ROOMS_COULD_NOT_BE_FETCHED = 'Rooms could not be fetched',
	ROOM_COULD_NOT_BE_FOUND = 'Room could not be found',
	ROOM_COULD_NOT_BE_CREATED = 'Room could not be created',
	ROOM_COULD_NOT_BE_UPDATED = 'Room could not be updated',
	ROOM_COULD_NOT_BE_DELETED = 'Room could not be deleted',
	ROOM_WITH_NAME_ALREADY_EXISTS = 'Room with name already exists',
	ROOM_ALREADY_EXISTS = 'Room already exists',
}